requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
tensorflow
keras
numpy
joblib
dvc
pandas
matplotlib
scipy
fastapi
uvicorn
pillow
numpy
